filecopy: do not use non-blocking fds, use select() instead
The right way to check if there is data to read is to select() on the
FD, not set to non-block and read() on it. Setting O_NONBLOCK flag on fd
can have side effects, for example when stdin/stdout uses the same
socket, this flag affects both (so writes can fail with EAGAIN).